move Diagnostic and the Logger module to compile separately
