+++ src/compiler/diagnostic/index.ts
@@ -1,0 +1,20 @@
+export interface Diagnostic {
+  level: 'error' | 'warn' | 'info' | 'log' | 'debug';
+  type: string;
+  header?: string;
+  language?: string;
+  messageText: string;
+  debugText?: string;
+  code?: string;
+  absFilePath?: string;
+  relFilePath?: string;
+  lineNumber?: number;
+  columnNumber?: number;
+  lines?: {
+    lineIndex: number;
+    lineNumber: number;
+    text?: string;
+    errorCharStart: number;
+    errorLength?: number;
+  }[];
+}
