+++ src/compiler/diagnostic/tsconfig.json
@@ -1,0 +1,10 @@
+{ 
+  "extends": "../../../tsconfig-base.json",
+  "compilerOptions": {
+    "composite": true,
+    "strict": true
+  },
+  "include": [
+    "./index.ts"
+  ]
+}
